+++ Sources/KeyboardKit/UIKit/Autocomplete/UIAutocompleteToolbarSeparator.swift
@@ -1,5 +1,5 @@
 //
-//  AutocompleteToolbarSeparator.swift
+//  UIAutocompleteToolbarSeparator.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2019-12-09.
@@ -11,11 +11,8 @@ import UIKit
 /**
  This view is added as `AutocompleteToolbarLabel` separators
  to separate autocomplete labels in an `AutocompleteToolbar`.
- 
- You can style the appearance of these separators, using the
- `AutocompleteToolbarSeparator.appearance` proxy.
  */
-public class AutocompleteToolbarSeparator: UIView {
+public class UIAutocompleteToolbarSeparator: UIView {
     
     public override func willMove(toWindow window: UIWindow?) {
         backgroundColor = .lightGray